feat(hub): created billing summary components
ref: MANAGER-15061
