--- packages/manager/apps/hub/src/components/enterprise-billing-summary/EnterpriseBillingSummary.constants.ts
@@ -1,1 +1,0 @@
-export const LINK = 'https://billing.us.ovhcloud.com/login';
