+++ packages/manager/apps/hub/src/hooks/periodFilter/usePeriodFilter.tsx
@@ -1,0 +1,14 @@
+import { Period } from '@/types/bills.type';
+
+export const usePeriodFilter = ({ from, to }: Period) => [
+  {
+    field: 'date',
+    comparator: 'isAfter',
+    reference: [from],
+  },
+  {
+    field: 'date',
+    comparator: 'isBefore',
+    reference: [to],
+  },
+];
