+++ packages/manager/apps/hub/src/pages/layout/layout.constants.ts
@@ -6,6 +6,12 @@ export const features = [
   'procedures:fraud',
 ];
 
+export const BILLING_SUMMARY_PERIODS_IN_MONTHS = [1, 3, 6];
+
+export const LINK = 'https://billing.us.ovhcloud.com/login';
+
 export default {
   features,
+  BILLING_SUMMARY_PERIODS_IN_MONTHS,
+  LINK,
 };
